Move the `validateGitHubWebHook` function into its own file
That way, things are organized better, and easier to test locally, too.

This commit is best viewed with `--color-moved`.
